fix(cpp): address clang-tidy and static analysis findings
- Remove unused intEncoder parameter from encodeMortonCodes
- Remove redundant uint32_t casts where variable is already uint32_t
- Add parentheses for operator precedence clarity in stream count math
- Use designated initializers for Vertex aggregate initialization
- Add build-asan/ to .gitignore

Remaining suppressed findings are project-wide patterns:
- performance-trivially-destructible: PIMPL requires out-of-line dtor
- llvm-prefer-static-over-anonymous-namespace: codebase uses anon ns
- portability-avoid-pragma-once: used consistently throughout project

Validated with: clang-tidy 21, cppcheck 2.17, ASan+UBSan (62/62 tests)

+++ cpp/src/mlt/encode/int.cpp
@@ -151,7 +151,7 @@ IntegerEncodingResult IntegerEncoder::encodeInt(std::span<const std::int32_t> va
         const auto rlePhysicalLength = static_cast<std::uint32_t>(rle.runs.size() + rle.values.size());
         if (isConstStream || rleEncoded.size() < best.data->size()) {
             best = {LogicalLevelTechnique::RLE, LogicalLevelTechnique::NONE,
-                    &rleEncoded, static_cast<std::uint32_t>(runs), rlePhysicalLength};
+                    &rleEncoded, runs, rlePhysicalLength};
         }
     }
 
@@ -170,7 +170,7 @@ IntegerEncodingResult IntegerEncoder::encodeInt(std::span<const std::int32_t> va
         const auto drlePhysicalLength = static_cast<std::uint32_t>(deltaRle.runs.size() + deltaRle.values.size());
         if (deltaRleEncoded.size() < best.data->size()) {
             best = {LogicalLevelTechnique::DELTA, LogicalLevelTechnique::RLE,
-                    &deltaRleEncoded, static_cast<std::uint32_t>(deltaRuns), drlePhysicalLength};
+                    &deltaRleEncoded, deltaRuns, drlePhysicalLength};
         }
     }
 
@@ -239,7 +239,7 @@ IntegerEncodingResult IntegerEncoder::encodeLong(std::span<const std::int64_t> v
         const auto rlePhysicalLength = static_cast<std::uint32_t>(rle.runs.size() + rle.values.size());
         if (rleEncoded.size() < best.data->size()) {
             best = {LogicalLevelTechnique::RLE, LogicalLevelTechnique::NONE,
-                    &rleEncoded, static_cast<std::uint32_t>(runs), rlePhysicalLength};
+                    &rleEncoded, runs, rlePhysicalLength};
         }
     }
 
@@ -260,7 +260,7 @@ IntegerEncodingResult IntegerEncoder::encodeLong(std::span<const std::int64_t> v
         const auto drlePhysicalLength = static_cast<std::uint32_t>(deltaRle.runs.size() + deltaRle.values.size());
         if (deltaRleEncoded.size() < best.data->size()) {
             best = {LogicalLevelTechnique::DELTA, LogicalLevelTechnique::RLE,
-                    &deltaRleEncoded, static_cast<std::uint32_t>(deltaRuns), drlePhysicalLength};
+                    &deltaRleEncoded, deltaRuns, drlePhysicalLength};
         }
     }
